third_party/blink/web_tests/external/wpt/css/css-transitions/parsing/starting-style-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] @starting-style is valid
  assert_equals: expected "@starting-style {\\n}" but got "@starting-style  {\\n}"
Harness: the test ran to completion.
